用户只能删除自己的资源，即给了指定资源id也要满足 =user_id

+++ dao/sqlmapping/ArticleSqlMapping.js
@@ -3,7 +3,7 @@
 const article = {
     insert:'INSERT INTO article(id, user_id, title, content, read_count) VALUES(0,?,?,?,?)',
     update:'update article set title=?, content=? where id=?',
-    remove: 'delete from article where id=?',
+    remove: 'delete from article where id=? and user_id=?',
     queryAll: 'select * from article_view limit ?, ?',
     queryById: 'select * from article_view where id=?',
     queryByTitle: 'select * from article_view where title like "%"?"%" limit ?, ?',
